chore(channel): Move BoxFuture to channel module
